requirements.txt
ultralytics
Pillow
opencv-python-headless
redis
python-dotenv